Refactoring + relocalisation des fichiers js et des services dans des fichiers approprié
